// src/domain/chart/enums.ts
export enum Planet {
  SUN = 'SUN',
  MOON = 'MOON',
  MERCURY = 'MERCURY',
  VENUS = 'VENUS',
  MARS = 'MARS',
  JUPITER = 'JUPITER',
  SATURN = 'SATURN',
  URANUS = 'URANUS',
  NEPTUNE = 'NEPTUNE',
  PLUTO = 'PLUTO',
  CHIRON = 'CHIRON',
}

export enum ZodiacSign {
  ARIES = 'ARIES',
  TAURUS = 'TAURUS',
  GEMINI = 'GEMINI',
  CANCER = 'CANCER',
  LEO = 'LEO',
  VIRGO = 'VIRGO',
  LIBRA = 'LIBRA',
  SCORPIO = 'SCORPIO',
  SAGITTARIUS = 'SAGITTARIUS',
  CAPRICORN = 'CAPRICORN',
  AQUARIUS = 'AQUARIUS',
  PISCES = 'PISCES',
}

export enum AspectType {
  CONJUNCTION = 'CONJUNCTION',
  SEXTILE = 'SEXTILE',
  SQUARE = 'SQUARE',
  TRINE = 'TRINE',
  OPPOSITION = 'OPPOSITION',
}

export enum HouseSystem {
  PLACIDUS = 'PLACIDUS',
}

export enum ZodiacType {
  TROPICAL = 'TROPICAL',
}

export enum AdditionalPointType {
  NORTH_NODE = 'NORTH_NODE',
  SOUTH_NODE = 'SOUTH_NODE',
  PART_OF_FORTUNE = 'PART_OF_FORTUNE',
  LILITH = 'LILITH',
}